Add actions and refactor
